ssh: Fix compile error on 64-bit systems.

+++ lib/ssh.c
@@ -2745,7 +2745,7 @@ static CURLcode ssh_block_statemach(struct connectdata *conn,
         fd_write = sock;
       /* wait for the socket to become ready */
       Curl_socket_ready(fd_read, fd_write,
-                        left>1000?1000:left); /* ignore result */
+                        (int)(left>1000?1000:left)); /* ignore result */
     }
 #endif
